Order of magnitude values are now stored as ints.

+++ MESolver.cpp
@@ -81,9 +81,9 @@ int main(int argc, char* argv[])
 
     gsl_matrix* A = gsl_matrix_alloc(M, M);
     gsl_matrix_set_zero(A);
-    double highestO = -999;
-    double lowestO = 999;
-    double orderOfMag;
+    int highestO = -999.0;
+    int lowestO = 999.0;
+    int orderOfMag;
     for (int i = 0; i < M; i++)
         for (int f = 0; f < M; f++)
         {
